Add `TrayProvider` to provide system tray services
`TrayProvider` provides the tray services. The tray is displayed in
the system notification area. add(...) and remove() functions are
available to the `main` process.
