+++ source/app/service-providers/assets/types.tray-provider.d.ts
@@ -1,0 +1,4 @@
+interface TrayProvider {
+  add: (show: () => void, quit: () => void) => void
+  remove: () => void
+}
